[HGS] Finally delete the old simulator
